+++ changelog.md
@@ -16,6 +16,7 @@ Last change: 25-SEP-2026 CEST
 - The display settings (main and 3D window) show the minimum blur only for the Gaussian blur methods that use it.
 - New action to move xy positions of localizations with a mouse. See the [documentation](https://picassosr.readthedocs.io/en/latest/render.html#move-ctrl-g).
 - Apply expression to localizations expands the canvas (metadata's `Height` and `Width`) if x and y positions are out of range.
+- Overlay of image files (`.png` and `.tif`), see the [documentation](https://picassosr.readthedocs.io/en/latest/render.html#overlay-image).
 - Faster circular picking of localizations.
 - Faster FRC.
 - FRC in several random ROIs (*Info* dialog, *FRC in several ROIs*).